+++ NEWS.md
@@ -1,3 +1,10 @@
+# verywise 1.3.8
+
+* Users can now specify `save_cov` argument to save the vertex-wise covariance between two beta estimates. This is useful for plotting and probing interactions.
+* Added a `<hemi>.<measure>.model.summary.yml` output for quickly inspecting results and model details that were previously only available in the log. 
+* Included more flexible cut-off logic to inspect difference maps
+* `se` of non-fitted models are now set to NA for consistency with beta values. 
+
 # verywise 1.3.7
 
 # verywise 1.3.3